# server/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/album": {
      "get": {
        "operationId": "getAllAlbums",
        "parameters": [
          {
            "name": "shared",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "assetId",
            "required": false,
            "in": "query",
            "description": "Only returns albums that contain the asset\nIgnores the shared parameter\nundefined: get all albums",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "operationId": "createAlbum",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/album/{id}": {
      "patch": {
        "operationId": "updateAlbumInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "get": {
        "operationId": "getAlbumInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/album/{id}/users": {
      "put": {
        "operationId": "addUsersToAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUsersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/album/{id}/user/{userId}": {
      "delete": {
        "operationId": "removeUserFromAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Album"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/share": {
      "get": {
        "operationId": "getAllSharedLinks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SharedLinkResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "share"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/share/me": {
      "get": {
        "operationId": "getMySharedLink",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "share"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/share/{id}": {
      "get": {
        "operationId": "getSharedLinkById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "share"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "patch": {
        "operationId": "updateSharedLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSharedLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedLinkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "share"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "operationId": "removeSharedLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "share"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.60.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "storageLabel": {
            "type": "string",
            "nullable": true
          },
          "profileImagePath": {
            "type": "string"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "oauthId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "storageLabel",
          "profileImagePath",
          "shouldChangePassword",
          "isAdmin",
          "createdAt",
          "deletedAt",
          "updatedAt",
          "oauthId"
        ]
      },
      "AssetTypeEnum": {
        "type": "string",
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ]
      },
      "ExifResponseDto": {
        "type": "object",
        "properties": {
          "fileSizeInByte": {
            "type": "integer",
            "nullable": true,
            "default": null,
            "format": "int64"
          },
          "make": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "model": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "exifImageWidth": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "exifImageHeight": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "orientation": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "dateTimeOriginal": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "default": null
          },
          "modifyDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "default": null
          },
          "timeZone": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "lensModel": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "fNumber": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "focalLength": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "iso": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "exposureTime": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "default": null
          },
          "city": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "state": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "country": {
            "type": "string",
            "nullable": true,
            "default": null
          },
          "description": {
            "type": "string",
            "nullable": true,
            "default": null
          }
        }
      },
      "SmartInfoResponseDto": {
        "type": "object",
        "properties": {
          "tags": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "objects": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TagTypeEnum": {
        "type": "string",
        "enum": [
          "OBJECT",
          "FACE",
          "CUSTOM"
        ]
      },
      "TagResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/TagTypeEnum"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "name",
          "userId"
        ]
      },
      "PersonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "thumbnailPath"
        ]
      },
      "AssetResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/AssetTypeEnum"
          },
          "id": {
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "resized": {
            "type": "boolean"
          },
          "thumbhash": {
            "type": "string",
            "nullable": true,
            "description": "base64 encoded thumbhash"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isArchived": {
            "type": "boolean"
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "smartInfo": {
            "$ref": "#/components/schemas/SmartInfoResponseDto"
          },
          "livePhotoVideoId": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            }
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonResponseDto"
            }
          },
          "checksum": {
            "type": "string",
            "description": "base64 encoded sha1 hash"
          }
        },
        "required": [
          "type",
          "id",
          "deviceAssetId",
          "ownerId",
          "deviceId",
          "originalPath",
          "originalFileName",
          "resized",
          "thumbhash",
          "fileCreatedAt",
          "fileModifiedAt",
          "updatedAt",
          "isFavorite",
          "isArchived",
          "mimeType",
          "duration",
          "checksum"
        ]
      },
      "AlbumResponseDto": {
        "type": "object",
        "properties": {
          "assetCount": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "albumName": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "albumThumbnailAssetId": {
            "type": "string",
            "nullable": true
          },
          "shared": {
            "type": "boolean"
          },
          "sharedUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            }
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "assetCount",
          "id",
          "ownerId",
          "albumName",
          "createdAt",
          "updatedAt",
          "albumThumbnailAssetId",
          "shared",
          "sharedUsers",
          "assets",
          "owner"
        ]
      },
      "CreateAlbumDto": {
        "type": "object",
        "properties": {
          "albumName": {
            "type": "string"
          },
          "sharedWithUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "assetIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "albumName"
        ]
      },
      "UpdateAlbumDto": {
        "type": "object",
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumThumbnailAssetId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AddUsersDto": {
        "type": "object",
        "properties": {
          "sharedUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "sharedUserIds"
        ]
      },
...
      "SharedLinkType": {
        "type": "string",
        "enum": [
          "ALBUM",
          "INDIVIDUAL"
        ]
      },
      "SharedLinkResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/SharedLinkType"
          },
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            }
          },
          "album": {
            "$ref": "#/components/schemas/AlbumResponseDto"
          },
          "allowUpload": {
            "type": "boolean"
          },
          "allowDownload": {
            "type": "boolean"
          },
          "showExif": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "id",
          "userId",
          "key",
          "createdAt",
          "expiresAt",
          "assets",
          "allowUpload",
          "allowDownload",
          "showExif"
        ]
      },
      "EditSharedLinkDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "allowUpload": {
            "type": "boolean"
          },
          "allowDownload": {
            "type": "boolean"
          },
          "showExif": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}